move shared function out of ontology chip file
