+++ libs/@hashintel/design-system/src/ontology-chip.tsx
@@ -1,8 +1,4 @@
 import {
-  validateVersionedUrl,
-  VersionedUrl,
-} from "@blockprotocol/type-system/slim";
-import {
   Box,
   Stack,
   SxProps,
@@ -13,22 +9,6 @@ import {
 import { forwardRef, ForwardRefRenderFunction, useMemo } from "react";
 
 import { IconRainbowHash } from "./icon-rainbow-hash";
-
-export const parseUrlForOntologyChip = (url: VersionedUrl) => {
-  const validationResult = validateVersionedUrl(url);
-  if (validationResult.type === "Err") {
-    throw new Error(
-      `Could not validate url as VersionedUrl: ${validationResult.inner.reason}`,
-    );
-  }
-  const parsed = validationResult.inner;
-  const parsedUrl = new URL(parsed);
-  const domain =
-    parsedUrl.host === "localhost:3000" ? "localhost" : parsedUrl.host;
-  const path = parsedUrl.pathname.slice(1);
-
-  return { domain, path };
-};
 
 // @todo make this take the id
 const OntologyChip: ForwardRefRenderFunction<
