+++ libs/@hashintel/design-system/src/parse-url-for-ontology-chip.tsx
@@ -1,0 +1,20 @@
+import {
+  validateVersionedUrl,
+  VersionedUrl,
+} from "@blockprotocol/type-system/slim";
+
+export const parseUrlForOntologyChip = (url: VersionedUrl) => {
+  const validationResult = validateVersionedUrl(url);
+  if (validationResult.type === "Err") {
+    throw new Error(
+      `Could not validate url as VersionedUrl: ${validationResult.inner.reason}`,
+    );
+  }
+  const parsed = validationResult.inner;
+  const parsedUrl = new URL(parsed);
+  const domain =
+    parsedUrl.host === "localhost:3000" ? "localhost" : parsedUrl.host;
+  const path = parsedUrl.pathname.slice(1);
+
+  return { domain, path };
+};
